[AIRFLOW-2886] Secure Flask SECRET_KEY
The Flask SECRET_KEY should be as random as possible.

On the other hand, we can't genrate random value when we
launch the webserver (the secret_key will be inconsistent
across the workers).

We can generate a random one in the configuration file
airflow.cfg, just like how we deal with FERNET_KEY.

The SECRET_KEY is generated using os.urandom, as
recommended by Flask community.